Refactor PR
