+++ apps/api/.env
@@ -6,10 +6,10 @@
 
 NODE_ENV=development
 
-DATABASE_URL=postgresql://dumpovac:dump123@localhost:3001/postgre?schema=public
+DATABASE_URL=postgresql://dumpovac:dump123@localhost:3001/internship-app
 
 JWT_SECRET=test1950
 
 AWS_ACCESS_KEY_ID=
 AWS_SECRET_ACCESS_KEY=
-#POSTMARK_API_TOKEN=
+POSTMARK_API_TOKEN=
