+++ apps/api/src/interview-question/interview-question.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 
 import { PrismaService } from '../prisma.service';
-import { QuestionController } from './question.controller';
-import { QuestionService } from './question.service';
+import { QuestionController } from './interview-question.controller';
+import { QuestionService } from './interview-question.service';
 
 @Module({
   controllers: [QuestionController],
